Pull out opt and type alias
