+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/mod.rs
@@ -3,13 +3,13 @@ use swc_ecma_ast::TsTypeAliasDecl;
 
 use crate::ts_ast::{source_map::SourceMapped, TsAst};
 
-mod ts_type_alias_decl;
-
 pub mod azle_type_ref;
 pub mod func;
+pub mod opt;
 pub mod record;
 pub mod service;
 pub mod tuple;
+pub mod type_alias;
 pub mod variant;
 
 impl TsAst {
