+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/opt.rs
@@ -1,0 +1,12 @@
+use cdk_framework::act::node::candid::Opt;
+
+use super::azle_type_ref::AzleTypeRef;
+
+impl AzleTypeRef<'_> {
+    pub fn to_option(&self) -> Opt {
+        let enclosed_act_data_type = self.get_enclosed_azle_type().to_data_type();
+        Opt {
+            enclosed_type: Box::from(enclosed_act_data_type),
+        }
+    }
+}
